requirements.txt
streamlit
pandas==2.0.3
pdfplumber
pytesseract==0.3.10
pdf2image==1.16.3
google-generativeai
dotenv
python-dotenv==1.0.0
PyPDF2==3.0.1
opencv-python==4.8.1.78
numpy==1.24.3
scikit-learn==1.3.0
spacy==3.6.1
python-magic==0.4.27
Pillow==10.0.0
tqdm==4.66.1
pytest==7.4.0
black==23.7.0
flake8==6.1.0
